requirements.txt
python-dotenv==1.0.1
openai==1.65.1
Flask==3.0.3
requests==2.32.3
redis==5.2.1
pydantic==2.10.3
loguru==0.7.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
supabase~=2.15.1
PyJWT~=2.10.1
httpx~=0.28.1